+++ src/Http/Controllers/PaymentCancelledController.php
@@ -1,0 +1,14 @@
+<?php
+
+namespace Naykel\Payit\Http\Controllers;
+
+class PaymentCancelledController
+{
+    public function __invoke()
+    {
+        session()->remove('payment');
+
+        return redirect()->route('checkout')
+            ->withErrors('You cancelled the payment.');
+    }
+}