+++ CHANGELOG.md
@@ -1,13 +1,13 @@
 # Changelog
 
-- Removed migration, factory and seeder (use sushi instead)
 - Remove PaymentController and replaced with single action controllers
-  
+
+
 ## 0.3.0
 - Added: `PaymentService` abstract class and updated services to extend it
 - Added `payit.php` config file and updated services to use it
 - Removed icons components (will be replaced with GT icons)
-
+- Removed migration, factory and seeder (use sushi instead)
 
 ## 0.2.0
 - Moved: controllers to http directory